Update input.xml
I don't think that you need the <type>. You know from the <name> tag which code variable you are going to store the parameter value in and so the type is predetermined. In .Net we use the Double.TryParse(sValue, fValue) or Integer.TryParse(sValue, nValue) to consume numeric values from text XML files.