Update results.xml
Keep the params and results sections separate. Results, after all, are not parameters. Also, reusing the structure of the parameters from the input XML file will make it easier on .Net to reuse the same code classes to read the result XML file as it does to write the input file.